fastapi_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tortoise-orm==0.20.1
pydantic==2.5.0
python-multipart==0.0.6
